A slight modification to RKL_REGISTER_FOR_IPHONE_LOWMEM_NOTIFICATIONS.
An attempt to enable the functionality of RKL_REGISTER_FOR_IPHONE_LOWMEM_NOTIFICATIONS is attempted in RegexKitLite.h by checking for TARGET_OS_EMBEDDED || TARGET_OS_IPHONE.  If those are enabled, and RKL_REGISTER_FOR_IPHONE_LOWMEM_NOTIFICATIONS is not defined or RKL_REGISTER_FOR_IPHONE_LOWMEM_NOTIFICATIONS != 0 (to allow explicitly disabling the functionality), then enable the feature. 

Hopefully, this allows 'the right thing' to be selected automatically in most cases.  Again, it's harmless if this is enabled 'by accident' since the notification that we register for is looked up dynamically via dlsym().  So, for example, if RKL_REGISTER_FOR_IPHONE_LOWMEM_NOTIFICATIONS is defined but built for standard Mac OS X, nothing (bad) happens. At load time we try to find the iPhone low mem notification symbol, but can't because it doesn't exist, and continue on our merry way.